Updating based on latest fix in core.
