+++ api/admin/controller/catalog_services.py
@@ -23,10 +23,9 @@ class CatalogServicesController(SettingsController):
         super(CatalogServicesController, self).__init__(manager)
         service_apis = [MARCExporter]
         self.protocols = self._get_integration_protocols(service_apis, protocol_name_attr="NAME")
-        mirror_integration_setting = self._mirror_integration_setting()
-        if mirror_integration_setting:
-            mirror_integration_setting["description"] = MARCExporter.SETTING_DESCRIPTION
-            self.protocols[0]['settings'].append(mirror_integration_setting)
+        self.protocols[0]['settings'].append(
+            MARCExporter.get_storage_settings(self._db)
+        )
 
     def process_catalog_services(self):
         self.require_system_admin()
